Refactoring properly my code
