+++ package.json
@@ -12,7 +12,10 @@
   },
   "dependencies": {
     "@heroicons/react": "^2.0.13",
+    "@types/react-transition-group": "^4.4.5",
     "dotenv": "^16.0.3",
+    "install": "^0.13.0",
+    "npm": "^9.3.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-router-dom": "^6.4.5",
