+++ resources/utils/classNames.ts
@@ -1,0 +1,3 @@
+export const classNames = (...classes: (false | null | undefined | string)[]): string => {
+    return classes.filter(Boolean).join(' ');
+};
